refactor: divide the Main component into server and client
