--- src/components/ErrorTestButton/ErrorTestButton.module.css
@@ -1,32 +1,0 @@
-.button {
-  display: block;
-  margin: 0 auto;
-  padding: 10px 20px;
-  background-color: #4caf50;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  width: 50%;
-  transition:
-    width 0.3s,
-    padding 0.3s;
-}
-
-.button:hover {
-  background-color: #45a049;
-}
-
-@media (max-width: 768px) {
-  .button {
-    width: 70%;
-    padding: 8px 16px;
-  }
-}
-
-@media (max-width: 480px) {
-  .button {
-    width: 90%;
-    padding: 6px 12px;
-  }
-}
