--- src/components/ErrorTestButton/ErrorTestButton.tsx
@@ -1,18 +1,0 @@
-'use client';
-
-import * as React from 'react';
-import styles from './ErrorTestButton.module.css';
-
-const ErrorTestButton: React.FC = () => {
-  const handleError = () => {
-    throw new Error('This is a test error');
-  };
-
-  return (
-    <button className={styles.button} onClick={handleError}>
-      Throw Error
-    </button>
-  );
-};
-
-export default ErrorTestButton;
